Refactor codegen logic
